// src/enums/ButtonCssClasses.ts
export enum ButtomCssClasses {
  NUMERICCALCBUTTON = 'NumericCalcButton',
  SIMPLEBUTTON = 'SimpleButton',
  DOUBLEBUTTON = 'DoubleButton',
  YELLOWOPERATIONCALCBUTTON = 'YellowOperationCalcButton',
  GRAYOPERATIONCALCBUTTON = 'GrayOperationCalcButton',
  DOUBLESIZEBUTTON13 = 'DoubleSizeButton13',
  DOUBLESIZEBUTTON24 = 'DoubleSizeButton24',
  DOUBLESIZEBUTTON35 = 'DoubleSizeButton35',
}